% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue-adjust.R
\name{adjust_pvalue}
\alias{adjust_pvalue}
\title{Adjust p values}
\usage{
adjust_pvalue(x, .FUN = "p.adjust", method = "holm", proc = "Holm", ...)
}
\arguments{
\item{x}{a correlate or md_tbl object.}

\item{.FUN}{adjust function, such as \code{stats::p.adjust()},
\code{multtest::mt.rawp2adjp()}.}

\item{method}{adjust method when .FUN is "p.adjust".}

\item{proc}{adjust method when .FUN is "mt.rawp2adjp".}

\item{...}{other parameters passing on to adjust function.}
}
\value{
a object same as x.
}
\description{
Adjust correlation p values based on user-specified method.
}
\author{
Hou Yun
}
