% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anno-link.R
\name{anno_link}
\alias{anno_link}
\title{Link Annotation}
\usage{
anno_link(
  data,
  mapping = NULL,
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  nudge_x = 0.1,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{aesthetic mappings parameters.}

\item{label.size, label.colour, label.family, label.fontface}{parameters for label.}

\item{nudge_x}{horizonal justification of label.}

\item{...}{extra parameters passing to layer function.}
}
\value{
a ggplot layer.
}
\description{
Link annotation
}
\author{
Hou Yun
}
