% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{as_md_tbl}
\alias{as_md_tbl}
\alias{as_md_tbl.matrix_data}
\alias{as_md_tbl.pro_tbl}
\alias{as_md_tbl.mantel_tbl}
\alias{as_md_tbl.easycorrelation}
\alias{as_md_tbl.correlate}
\alias{as_md_tbl.rcorr}
\alias{as_md_tbl.corr.test}
\alias{as_md_tbl.data.frame}
\alias{as_md_tbl.matrix}
\title{Coerce matrix_data to data frames}
\usage{
as_md_tbl(x)

\method{as_md_tbl}{matrix_data}(x, ...)

\method{as_md_tbl}{pro_tbl}(x, byrow = TRUE, ...)

\method{as_md_tbl}{mantel_tbl}(x, byrow = TRUE, ...)

\method{as_md_tbl}{easycorrelation}(x, type = "full", diag = TRUE, ...)

\method{as_md_tbl}{correlate}(x, ...)

\method{as_md_tbl}{rcorr}(x, ...)

\method{as_md_tbl}{corr.test}(x, ...)

\method{as_md_tbl}{data.frame}(x, name = NULL, ...)

\method{as_md_tbl}{matrix}(x, name = NULL, ...)
}
\arguments{
\item{x}{any \code{R} can be converted to md_tbl.}

\item{...}{other parameters.}

\item{byrow}{a logical value indicating whether arrange the 'spec' columns
on y axis.}

\item{type}{character, "full" (default), "upper" or "lower", display full
matrix, lower triangular or upper triangular matrix.}

\item{diag}{logical, if TRUE (default) will keep the diagonal of matrix data.}

\item{name}{variable name.}
}
\value{
a tibble object.
}
\description{
Coerce matrix_data to data frames
}
\author{
Hou Yun
}
