% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-tibble.R
\name{as_tibble.matrix_data}
\alias{as_tibble.matrix_data}
\title{Coerce matrix_data to data frames}
\usage{
\method{as_tibble}{matrix_data}(md)
}
\arguments{
\item{md}{a matrix_data object.}
}
\value{
a tibble object.
}
\description{
Coerce matrix_data to data frames
}
\author{
Hou Yun
}
