% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{secondary_x_axis}
\alias{secondary_x_axis}
\alias{secondary_y_axis}
\alias{set_secondary_axis}
\title{Secondary and Nested Axis}
\usage{
secondary_x_axis(...)

secondary_y_axis(...)

set_secondary_axis(..., position = waiver(), position_aes = "x")
}
\arguments{
\item{...}{other parameters passing to \code{ggh4x::guide_axis_manual()} or
\code{ggh4x::guide_axis_nested()}.}

\item{position}{where this guide should be drawn: one of top, bottom, left, or right.}

\item{position_aes}{one of "x" or "y".}
}
\value{
a guide object.
}
\description{
functions to add secondary and nested axis.
}
\examples{
\dontrun{
as_md_tbl(mtcars) \%>\%
hyplot() +
  geom_point() +
  secondary_x_axis()
}
}
\author{
Hou Yun
}
