% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doughnut-grob.R
\name{DoughnutGrob}
\alias{DoughnutGrob}
\alias{grid.doughnut}
\title{Draw Doughnut}
\usage{
DoughnutGrob(
  x = 0.5,
  y = 0.5,
  r0 = 4,
  r1 = 5,
  value = runif(5),
  label = NULL,
  label_size = 7.5,
  label_col = "black",
  units = "mm",
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL
)

grid.doughnut(...)
}
\arguments{
\item{x, y}{a numeric value or unit object.}

\item{r0, r1}{a non-negtive numeric value.}

\item{value}{a numeric vector.}

\item{label}{NULL or character vector with the same length as 'value'.}

\item{label_size}{size of labels.}

\item{label_col}{color of labels.}

\item{units}{unit of the piechart size.}

\item{default.units}{a string indicating the default units to use if x or y
are only given as numeric vectors.}

\item{name}{a character identifier.}

\item{gp}{an object of class 'gpar'.}

\item{vp}{a viewport object or NULL.}

\item{...}{other parameters for \code{DoughnutGrob()}.}
}
\value{
a doughnut grob.
}
\description{
These functions create and draw doughnut.
}
\author{
Hou Yun
}
