% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-shaping.R
\name{draw_key_marker}
\alias{draw_key_marker}
\title{Key glyphs for marker}
\usage{
draw_key_marker(data, params, size)
}
\arguments{
\item{data}{a data frame.}

\item{params}{a list of additional parameters.}

\item{size}{width/height of key in mm.}
}
\value{
a grid grob object.
}
\description{
Designed to customize marker legends.
}
\author{
Hou Yun
}
