% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify.matrix_data}
\alias{fortify.matrix_data}
\alias{fortify.md_tbl}
\title{Fortify a matrix_data}
\usage{
\method{fortify}{matrix_data}(model, data, ...)

\method{fortify}{md_tbl}(model, data, ...)
}
\arguments{
\item{model}{a matrix_data object.}

\item{data}{ignore.}

\item{...}{ignore.}
}
\value{
a tibble object.
}
\description{
Fortify a matrix_data to data.frame
}
\author{
Hou Yun
}
