% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdist.R
\name{gdist}
\alias{gdist}
\alias{dist_func}
\title{Distance Matrix Computation}
\usage{
gdist(.data, .FUN = NULL, .formula = NULL, ...)

dist_func(...)
}
\arguments{
\item{.data}{a data frame object.}

\item{.FUN}{the name of dist function.}

\item{.formula}{lambda function.}

\item{...}{other parameters passing to dist function.}
}
\value{
dist matrix.
}
\description{
This function can computes and returns the distance matrix
based on the data type.
}
\examples{
gdist(mtcars)
gdist(iris)

m <- matrix(sample(LETTERS, 100, TRUE), nrow = 10)
gdist(m)
}
\author{
Hou Yun
}
