% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-annotate.R
\name{geom_annotate}
\alias{geom_annotate}
\title{Annotate Layer}
\usage{
geom_annotate(
  annotate = NULL,
  ...,
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "rt",
  inherit.aes = TRUE,
  show.legend = FALSE,
  width = NULL,
  height = NULL,
  nudge_x = NULL,
  nudge_y = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{annotate}{a grob object, or other object can be converted to a grob.}

\item{...}{other parameters passing to convert-function.}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{string, position of annotate:
 \itemize{
   \item{\code{"rt"}: right-top corner.}
   \item{\code{"rb"}: right-bottom corner.}
   \item{\code{"rc"}: right-center.}
   \item{\code{"lt"}: left-top corner.}
   \item{\code{"lb"}: left-bottom corner.}
   \item{\code{"lc"}: left-center.}
   \item{\code{"cc"}: center-center.}
   \item{\code{"ct"}: center-top.}
   \item{\code{"cb"}: center-bottom.}
}}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{width, height}{width/height of annotate.}

\item{nudge_x, nudge_y}{a minor shift of position, should be a grid::unit object.}
}
\value{
a layer object.
}
\description{
A wrapper funtion to add some annotate elements on ggplot.
}
