% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-couple.R
\name{geom_couple}
\alias{geom_couple}
\alias{anno_link}
\title{Link Layer}
\usage{
geom_couple(
  data,
  mapping = NULL,
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  nudge_x = 0.5,
  ...
)

anno_link(...)
}
\arguments{
\item{data}{a data frame.}

\item{mapping}{aesthetic mappings parameters.}

\item{label.size, label.colour, label.family, label.fontface}{parameters for label.}

\item{nudge_x}{horizonal justification of label.}

\item{...}{extra parameters passing to layer function.}
}
\value{
a ggplot layer.
}
\description{
Function to draw mantel test plot on a correlation matrix heatmap.
}
\note{
`anno_link()` has been soft deprecated, please use `geom_couple()`
instead.
}
\author{
Hou Yun
}
