% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-diag-label.R
\name{geom_diag_label}
\alias{geom_diag_label}
\title{Add diagonal label on hyplot}
\usage{
geom_diag_label(..., parse = FALSE, geom = NULL)
}
\arguments{
\item{...}{extra parameters.}

\item{parse}{logical or function generated by \code{parse_func()}.}

\item{geom}{one of "text", "label" or "richtext" (needs `ggtext` package).}
}
\description{
\code{geom_panel_grid} is mainly used with \code{hyplot()}.
    function to add a panel grid line on plot region.
Add diagnoal labels on correlation plot

\code{geom_diag_label} is mainly used with \code{hyplot()} and
    \code{qcorrplot()} functions to add diagnoal labels on correct position
    base on different type of cor_tbl object.
}
\examples{
qcorrplot(correlate(mtcars)) + geom_diag_label()
}
\author{
Hou Yun
}
