% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-panel-grid.R
\name{geom_panel_grid}
\alias{geom_panel_grid}
\title{Add panel grid line on hyplot}
\usage{
geom_panel_grid(colour = "grey50", size = 0.25, color = NULL)
}
\arguments{
\item{colour, color}{colour of grid lines.}

\item{size}{size of grid lines.}
}
\description{
\code{geom_panel_grid} is mainly used with \code{hyplot()}.
    function to add a panel grid line on plot region.
}
