% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-matrix-data.R
\name{row_names}
\alias{row_names}
\alias{row_names<-}
\alias{col_names}
\alias{col_names<-}
\alias{nrows}
\alias{ncols}
\alias{is_matrix_data}
\alias{is_md_tbl}
\alias{extract_upper}
\alias{extract_lower}
\alias{extract_diag}
\alias{trim_diag}
\title{Helper function}
\usage{
row_names(md)

row_names(md) <- value

col_names(md)

col_names(md) <- value

nrows(md)

ncols(md)

is_matrix_data(md)

is_md_tbl(md)

extract_upper(md, diag = TRUE)

extract_lower(md, diag = TRUE)

extract_diag(md)

trim_diag(md)
}
\arguments{
\item{md}{a matrix_data object.}

\item{value}{a valid value for dimension names.}

\item{diag}{logical, if TRUE (default) will keep the diagonal of matrix data.}
}
\description{
Helper function for matrix_data object
}
\author{
Hou Yun
}
