% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-utils.R
\name{label_formula}
\alias{label_formula}
\title{Label formula}
\usage{
label_formula(parse = TRUE)
}
\arguments{
\item{parse}{logical or function generated by \code{parse_func()}.}
}
\value{
a label function.
}
\description{
\code{label_formula()} can be used to parse text label to expression
or richtext, and displays labels as math formula.
}
\author{
Hou Yun
}
