% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-magic-text.R
\name{latex_richtext}
\alias{latex_richtext}
\title{Converts a LaTeX String to Richtext}
\usage{
latex_richtext(x, sup = "^", sub = "_", br = "\\n", env = parent.frame())
}
\arguments{
\item{x}{a character vector.}

\item{sup}{one-length character, indicates that characters after this is superscript.}

\item{sub}{one-length character, indicates that characters after this is subscript.}

\item{br}{string,  separator of lines.}

\item{env}{environment to evaluate each expression in.}
}
\description{
Helper function to convert a LaTeX string to richtext.
}
\examples{
name <- c("A_2", "B^3", "C_2", "D^{123 + x}")
name <- latex_richtext(name)
}
\author{
Hou Yun
}
