% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_with_circular}
\alias{layout_with_circular}
\alias{layout_tbl_graph_circular}
\title{Circular Layout}
\usage{
layout_with_circular(g, group = NULL, zoom = 0.618)

layout_tbl_graph_circular(g, group = NULL, zoom = 0.618, circular)
}
\arguments{
\item{g}{graph.}

\item{group}{if NULL (default), will use group attributes of `g`.}

\item{zoom}{numeric, range in [0, 1].}

\item{circular}{not used.}
}
\value{
a two-columns matrix.
}
\description{
arranges nodes in circles according to group information.
}
\author{
Hou Yun
}
