% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-annotate.R, R/geom-curve2.R,
%   R/geom-doughnut.R, R/geom-mark.R, R/geom-shaping.R, R/geom-square.R,
%   R/ggproto.R
\docType{data}
\name{GeomAnnotate}
\alias{GeomAnnotate}
\alias{GeomCurve2}
\alias{GeomDoughnut}
\alias{GeomMark}
\alias{Geomshaping}
\alias{GeomSquare}
\alias{linkET-extensions}
\title{linkET package extensions to ggplot2}
\description{
In general the actual classes should be of little interest to users as the
standard ggplot2 api of using geom_* and stat_* functions for building up
the plot is encouraged.
}
\keyword{datasets}
