% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-and-reorder.R
\name{make_cluster}
\alias{make_cluster}
\title{Hierarchical Clustering}
\usage{
make_cluster(
  md,
  index = 1,
  cluster = "none",
  cluster_method = "complete",
  row_dist = NULL,
  col_dist = NULL,
  ...
)
}
\arguments{
\item{md}{a matrix_data object.}

\item{index}{a integer or character index, specifying which matrix to cluster by.}

\item{cluster}{one of "all", "row", "col" or "none".}

\item{cluster_method}{the clustering agglomeration method to be used.}

\item{row_dist, col_dist}{dist function, see details.}

\item{...}{extra parameters for gdist.}
}
\value{
a modified matrix_data object.
}
\description{
Reorder matrix_data by hierarchical cluster
}
\author{
Hou Yun
}
