% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{makeContent.markerGrob}
\alias{makeContent.markerGrob}
\title{Dynamic resize the marker}
\usage{
\method{makeContent}{markerGrob}(x)
}
\arguments{
\item{x}{a marker grob object.}
}
\description{
This function takes care of updating the marker size based on
device size.
}
\author{
Hou Yun
}
