% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker.R
\name{marker}
\alias{marker}
\title{Custom marker for heatmap plot}
\usage{
marker(
  grob = "square",
  width = 1,
  height = 1,
  width_unit = "cm",
  height_unit = width_unit,
  r = 0,
  n = 5,
  ratio = 0.618
)
}
\arguments{
\item{grob}{grob object, list or character.
\itemize{
     \item{\strong{grob}: a \code{grob}, \code{gList}, \code{gTree} object.}
     \item{\strong{list}: a list of x and y.}
     \item{\strong{character}: one of "square", "circle", "star", "heart",
     "ellipse", "cross", "triangle", "triangle2".}
}}

\item{width, height}{width or height of marker.}

\item{width_unit, height_unit}{units of width or height.}

\item{r}{only be used for ellipse marker.}

\item{n, ratio}{only be used for star marker.}
}
\value{
a marker object.
}
\description{
This set of functions makes it easy to define shapes, and it
similar to points.
}
\examples{
marker()
marker(c("square", "circle"))
}
\author{
Hou Yun
}
