% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-data.R
\name{matrix_data}
\alias{matrix_data}
\title{Matrix data}
\usage{
matrix_data(
  x,
  type = "full",
  diag = TRUE,
  row_names = NULL,
  col_names = NULL,
  group = NULL,
  ...
)
}
\arguments{
\item{x}{a list of matrix-like data.}

\item{type}{character, "full" (default), "upper" or "lower", display full
matrix, lower triangular or upper triangular matrix.}

\item{diag}{logical, if TRUE (default) will keep the diagonal of matrix data.}

\item{row_names, col_names}{the name of rows and columns.}

\item{group}{NULL or a character vector.}

\item{...}{not used.}
}
\value{
a object of matrix_data
}
\description{
Matrix data
}
\author{
Hou Yun
}
