% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental-fun.R
\name{plot_type}
\alias{plot_type}
\title{Set Plot Type}
\usage{
plot_type(...)
}
\arguments{
\item{...}{arguments in \code{tag = value} form.}
}
\value{
a list of plot type.
}
\description{
This function can be used to set plot type of scatter matrix plot.
}
\author{
Hou Yun
}
