% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{print.correlate}
\alias{print.correlate}
\title{Print correlate object.}
\usage{
\method{print}{correlate}(x, all = FALSE, ...)
}
\arguments{
\item{x}{a correlate object.}

\item{all}{if FALSE (default) just print correlation matrix, else will
print all values.}

\item{...}{extra params passing to \code{print}.}
}
\description{
Print correlate object.
}
\examples{
m <- correlate(mtcars, cor.test = TRUE)
print(m)
print(m, TRUE)
}
\author{
Hou Yun
}
