% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcorrplot.R
\name{qcorrplot}
\alias{qcorrplot}
\alias{qcorrplot.correlate}
\alias{qcorrplot.rcorr}
\alias{qcorrplot.corr.test}
\alias{qcorrplot.mantel_tbl}
\alias{qcorrplot.pro_tbl}
\alias{qcorrplot.matrix}
\alias{qcorrplot.data.frame}
\title{Correlation Plot
These functions can be used to visualize simply and directly a correlation matrix.}
\usage{
qcorrplot(data, ...)

\method{qcorrplot}{correlate}(
  data,
  mapping = NULL,
  drop = FALSE,
  grid_col = "grey50",
  grid_size = 0.25,
  fixed = TRUE,
  ...
)

\method{qcorrplot}{rcorr}(
  data,
  mapping = NULL,
  drop = FALSE,
  grid_col = "grey50",
  grid_size = 0.25,
  fixed = TRUE,
  ...
)

\method{qcorrplot}{corr.test}(
  data,
  mapping = NULL,
  drop = FALSE,
  grid_col = "grey50",
  grid_size = 0.25,
  fixed = TRUE,
  ...
)

\method{qcorrplot}{mantel_tbl}(
  data,
  mapping = NULL,
  byrow = TRUE,
  drop = FALSE,
  grid_col = "grey50",
  grid_size = 0.25,
  fixed = TRUE,
  ...
)

\method{qcorrplot}{pro_tbl}(
  data,
  mapping = NULL,
  byrow = TRUE,
  drop = FALSE,
  grid_col = "grey50",
  grid_size = 0.25,
  fixed = TRUE,
  ...
)

\method{qcorrplot}{matrix}(
  data,
  mapping = NULL,
  drop = FALSE,
  grid_col = "grey50",
  grid_size = 0.25,
  is_corr = NULL,
  fixed = TRUE,
  ...
)

\method{qcorrplot}{data.frame}(
  data,
  mapping = NULL,
  drop = FALSE,
  grid_col = "grey50",
  grid_size = 0.25,
  is_corr = NULL,
  fixed = TRUE,
  ...
)
}
\arguments{
\item{data}{one of 'correlate', 'rcorr' or 'corr.test' object.}

\item{...}{passing to \code{\link{as_matrix_data}}.}

\item{mapping}{default list of aesthetic mappings to use for plot.}

\item{drop}{logical.}

\item{grid_col}{colour of panel grid.}

\item{grid_size}{size of panel grid.}

\item{fixed}{if TRUE (default), the coordinates will have a fixed aspect ratio.}

\item{byrow}{a logical value indicating whether arrange the 'spec' columns on y axis.}

\item{is_corr}{if TRUE (default), the data will be regarded as the correlation
coefficient matrix.}
}
\value{
a gg object.
}
\description{
Correlation Plot
These functions can be used to visualize simply and directly a correlation matrix.
}
\author{
Hou Yun
}
