% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlink.R
\name{qlink}
\alias{qlink}
\alias{extract_nodes}
\title{Draw link-curves}
\usage{
qlink(
  graph,
  mapping = NULL,
  ...,
  geom = "path",
  widths = NULL,
  heights = NULL,
  guides = "collect",
  nsteps = 100,
  hratio = 0.33
)

extract_nodes()
}
\arguments{
\item{graph}{a igraph, tbl_graph, or other can be converted to igraph object.}

\item{mapping}{default list of aesthetic mappings to use for plot.}

\item{...}{margin plot, should be named with "r", "l", "t" or "b".}

\item{geom}{no used.}

\item{widths}{width of margin plot.}

\item{heights}{height of margin plot.}

\item{guides}{a string specifying how guides should be treated in the layout.}

\item{nsteps}{integer, the value determines the smoothness of the curve.}

\item{hratio}{range 0 to 1, height ratio of half-circle link.}
}
\value{
a ggplot object.
}
\description{
This function can be used to draw a special network, which nodes
should be matched width margin plot.
}
\author{
Hou Yun
}
