% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest.R
\name{random_forest}
\alias{random_forest}
\alias{print.random_forest}
\title{Random forests}
\usage{
random_forest(spec, env, byrow = TRUE, seed = 123, ...)

\method{print}{random_forest}(x, ...)
}
\arguments{
\item{spec, env}{a data.frame object.}

\item{byrow}{a logical value, if TRUE, the 'spec' on the rows.}

\item{seed}{a integer value.}

\item{...}{extra parameters.}

\item{x}{a rand_forest object.}
}
\value{
a rand_forest object.
}
\description{
Random forests
}
\examples{
\dontrun{
spec <- mtcars[c(1, 3, 4, 5)]
env <- mtcars[6:11]
random_forest(spec, env)
}
}
\author{
Hou Yun
}
