% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{aes}
\alias{\%>\%}
\alias{as_tibble}
\alias{as.igraph}
\alias{as_tbl_graph}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{aes}}}

  \item{igraph}{\code{\link[igraph]{as.igraph}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}}

  \item{tidygraph}{\code{\link[tidygraph:tbl_graph]{as_tbl_graph}}}
}}

