% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental-fun.R
\name{register_pairs_plot}
\alias{register_pairs_plot}
\title{Register pairs plot}
\usage{
register_pairs_plot(...)
}
\arguments{
\item{...}{any valid layer parameters.}
}
\value{
set global options and return NULL.
}
\description{
Init pairs plot layer function.
}
\author{
Hou Yun
}
