% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder.R
\name{reorder_rows}
\alias{reorder_rows}
\alias{reorder_cols}
\title{Reorder Matrix Data}
\usage{
reorder_rows(x, by = NULL, ...)

reorder_cols(x, by = NULL, ...)
}
\arguments{
\item{x}{a \code{matrix_data} or \code{md_tbl} object.}

\item{by}{any \code{R} object.}

\item{...}{when \code{x} is a \code{matrix_data} object and \code{by} is NULL,
these are passed to the \code{gdist()} function.}
}
\value{
a modified \code{matrix_data} or \code{md_tbl} object.
}
\description{
reorder a \code{matrix_data} or \code{md_tbl} object.
}
\examples{
x <- as_matrix_data(mtcars)
reorder_rows(x)
}
\author{
Hou Yun
}
