% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental-fun.R
\name{set_pairs_scale}
\alias{set_pairs_scale}
\title{Set aesthestic scale}
\usage{
set_pairs_scale(..., ID = NULL)
}
\arguments{
\item{...}{scale object.}

\item{ID}{specify the ID of the element plot that you want to add.}
}
\value{
a ggplot.
}
\description{
Before drawing, you can set the scale function for each plot of pairs.
}
\author{
Hou Yun
}
