% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-hy.R
\name{theme_hy}
\alias{theme_hy}
\title{hyplot theme}
\usage{
theme_hy(..., use_md = FALSE)
}
\arguments{
\item{...}{extra params passing to \code{\link[ggplot2]{theme}}.}

\item{use_md}{logical. if TRUE, will use \code{ggtext::element_markdown()} to
draw the axis labels.}
}
\value{
The theme.
}
\description{
Create the default hyplot theme
}
\author{
Hou Yun
}
