% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap-annotate.R
\name{wrap_annotate}
\alias{wrap_annotate}
\title{Add Margin Annotation}
\usage{
wrap_annotate(
  plot,
  ...,
  width = NULL,
  height = NULL,
  k = 2,
  colour = NULL,
  border = NA,
  color = NULL
)
}
\arguments{
\item{plot}{a ggplot object.}

\item{...}{annotation elements.}

\item{width, height}{width/height of each annotation plot.}

\item{k}{an integer scalar with the desired number of groups.}

\item{border, colour, color}{fill and border colour of the hclust bar.}
}
\description{
This function is a wrapper of \code{aplot}, and can be used to
insert annotation.
}
\references{
Yu G (2022). aplot: Decorate a 'ggplot' with Associated Information. R package
version 0.1.6, https://CRAN.R-project.org/package=aplot.
}
\author{
Hou Yun
}
