% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{adjacency_matrix}
\alias{adjacency_matrix}
\title{Generate adjacency matrix}
\usage{
adjacency_matrix(x, ..., diag = FALSE)
}
\arguments{
\item{x}{a correlation object (\code{correlate}, \code{md_tbl}, and so on).}

\item{...}{expressions that return a logical value.}

\item{diag}{if FALSE (default),  self-edges will be removed.}
}
\value{
a matrix object.
}
\description{
Can be used to convert a correlation object to adjacency matrix.
}
\author{
Hou Yun
}
