% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue-adjust.R
\name{adjust_pvalue}
\alias{adjust_pvalue}
\title{Adjust p values}
\usage{
adjust_pvalue(x, method = "holm", ...)
}
\arguments{
\item{x}{a correlate or md_tbl object.}

\item{method}{adjust method.}

\item{...}{other parameters passing on to adjust function.}
}
\value{
a object same as x.
}
\description{
Adjust correlation p values based on user-specified method.
}
\author{
Hou Yun
}
