% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-pal.R
\name{color_pal}
\alias{color_pal}
\alias{colour_pal}
\title{Colour Palettes}
\usage{
color_pal(n, alpha = 0.6)

colour_pal(n, alpha = 0.6)
}
\arguments{
\item{n}{integer, should be less than 6.}

\item{alpha}{alpha level in [0,1].}
}
\value{
a palette.
}
\description{
Colour palettes for mantel test plot.
}
\examples{
color_pal(4)
}
\author{
Hou Yun
}
