% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Doughnut-grob.R
\name{DoughnutGrob}
\alias{DoughnutGrob}
\alias{grid.doughnut}
\title{Draw doughnut plot}
\usage{
DoughnutGrob(
  x = 0.5,
  y = 0.5,
  r0 = 0,
  r1 = 5,
  value = runif(5),
  percent = FALSE,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL
)

grid.doughnut(...)
}
\arguments{
\item{x, y}{a numeric value or unit object.}

\item{r0, r1}{a non-negtive numeric value.}

\item{value}{a numeric vector.}

\item{percent}{logical. If FALSE (the default) the value will be treated as
source value.}

\item{default.units}{a string indicating the default units to use if x or y
are only given as numeric vectors.}

\item{name}{a character identifier.}

\item{gp}{an object of class 'gpar'.}

\item{vp}{a viewport object or NULL.}

\item{...}{other parameters for \code{DoughnutGrob()}.}
}
\value{
a doughnut grob.
}
\description{
These functions create and draw doughnut.
}
\author{
Hou Yun
}
