% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/geom-couple.R
\name{anno_link}
\alias{anno_link}
\alias{geom_couple}
\alias{nice_curvature}
\title{Link Layer}
\usage{
anno_link(...)

geom_couple(
  data,
  mapping = NULL,
  drop = TRUE,
  label.size = 3.88,
  label.colour = "black",
  label.family = "",
  label.fontface = 1,
  nudge_x = 0.5,
  offset_x = NULL,
  offset_y = NULL,
  ...
)

nice_curvature(curvature, by = "to")
}
\arguments{
\item{...}{extra parameters passing to layer function.}

\item{data}{a data frame.}

\item{mapping}{aesthetic mappings parameters.}

\item{drop}{logical. If TRUE (default), the unused levels will be droped.}

\item{label.size, label.colour, label.family, label.fontface}{parameters for label.}

\item{nudge_x}{horizonal justification of label.}

\item{offset_x, offset_y}{NULL or named-list, add a little offset to the node points.}

\item{curvature}{a numeric value giving the amount of curvature.}

\item{by}{one of "from" or "to". Compute the best curvature based on
start or end points.}
}
\value{
a ggplot layer.
}
\description{
Function to draw mantel test plot on a correlation matrix heatmap.
}
\note{
`anno_link()` has been deprecated in 0.0.4, please use `geom_couple()`
instead.
}
\author{
Hou Yun
}
