% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-magic-text.R
\name{geom_magic_text}
\alias{geom_magic_text}
\title{Magic Method to Draw labels}
\usage{
geom_magic_text(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  geom = "text",
  parse = TRUE,
  ...
)
}
\arguments{
\item{mapping}{set of aesthetic mappings created by \code{aes()} or \code{aes_()}.}

\item{data}{the data to be displayed in this layer.}

\item{stat}{the statistical transformation to use on the data for this layer,
as a string.}

\item{position}{position adjustment, either as a string, or the result of a
call to a position adjustment function.}

\item{geom}{one of text, label or richtext.}

\item{parse}{logical or a parse function. IF TRUE (default), the labels will
be parsed into expression.}

\item{...}{others passing to \code{ggplot2::geom_<geom>()}.}
}
\value{
a gg layer object.
}
\description{
a wrapper function of \code{ggtext::geom_richtext}, but can
formatted labels by \code{LaTeX} style.
}
\author{
Hou Yun
}
