% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel-test.R
\name{regex_select}
\alias{regex_select}
\alias{prefix_with}
\alias{suffix_with}
\alias{contain_with}
\title{Subset columns using regular expression
Subset columns from a data frame if column names match the regular pattern.}
\usage{
regex_select(prefix = NULL, suffix = NULL, regex = NULL, ..., byrow = FALSE)

prefix_with(data, chr = "", ..., byrow = FALSE)

suffix_with(data, chr = "", ..., byrow = FALSE)

contain_with(data, chr = "", ..., byrow = FALSE)
}
\arguments{
\item{prefix}{prefix of column names.}

\item{suffix}{suffix of column names.}

\item{regex}{other string (or regular expression) in column names.}

\item{...}{other parameters passing to \code{grepl}.}

\item{byrow}{logical, if TRUE will select rows based on regex.}

\item{data}{a data frame.}

\item{chr}{a character used to match the column names.}
}
\value{
a data frame.
}
\description{
Subset columns using regular expression
Subset columns from a data frame if column names match the regular pattern.
}
\examples{
prefix_with(mtcars, "m")
}
\author{
Hou yun
}
