% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental-fun.R
\name{register_pairs_plot}
\alias{register_pairs_plot}
\title{Register pairs plot}
\usage{
register_pairs_plot(..., ptype = NULL, scale = list(), reset = FALSE)
}
\arguments{
\item{...}{any valid layer parameters.}

\item{ptype}{a plot_type object, which can be created by `plot_type()`.}

\item{scale}{a list of aesthestic scale.}

\item{reset}{logical, if TRUE will reset pairs plot parameters.}
}
\value{
set global options and return NULL.
}
\description{
Init pairs plot layer function.
}
\author{
Hou Yun
}
