% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-hy.R
\name{theme_hy}
\alias{theme_hy}
\title{hyplot theme}
\usage{
theme_hy(...)
}
\arguments{
\item{...}{extra params passing to \code{\link[ggplot2]{theme}}.}
}
\value{
The theme.
}
\description{
Create the default hyplot theme
}
\author{
Hou Yun
}
